+++ logmerger/about.py
@@ -1,4 +1,4 @@
-VERSION = "0.10.1"
+VERSION = "0.11.0"
 RELEASE_YEAR = "2025"
 
 text = fr"""
@@ -97,7 +97,7 @@ that use seconds-since-epoch timestamps more human-readable.
 - CSV files (timestamp is read from first data column)
   - (filename ending in `.csv`)
 - JSONL files - files containing a JSON object per-line (timestamp is read from first data column)
-  - (filename ending in `.jsonl`)\
+  - (filename ending in `.jsonl`)
   - if `orjson` package is installed, uses that for JSON parsing instead of the the stdlib json module.
 - packet capture files, created using tcpdump or WireShark (experimental)
   - (filename ending in `.pcap`)